♻️ refactor: 얼굴 인식 기능 컴포넌트 분리
