+++ src/pages/user/FaceRegistration.tsx
@@ -1,5 +1,12 @@
+import FaceDetection from '../../components/face/FaceDetection';
+
 const FaceRegistration = () => {
-  return <div>FaceRegistration</div>;
+  return (
+    <div>
+      FaceRegistration
+      <FaceDetection />
+    </div>
+  );
 };
 
 export default FaceRegistration;
